requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
dj-database-url==2.1.0
Django==4.2.13
django-cors-headers==4.3.1
django-filter==24.2
django-heroku==0.3.1
djangorestframework==3.15.1
filelock==3.14.0
flake8==7.0.0
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
importlib_metadata==7.1.0
Markdown==3.6
mccabe==0.7.0
packaging==23.2
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PySocks==1.7.1
requests==2.31.0
requests-file==2.0.0
sqlparse==0.5.0
tldextract==5.1.2
typing_extensions==4.11.0
urllib3==2.2.1
whitenoise==6.6.0
zipp==3.18.1
zope.event==5.0
zope.interface==6.3
django-oauth-toolkit
twilio
langchain==0.2.1
langchain-community 
langchainhub 
langchain-openai 
langchain-chroma 
bs4
langchain-anthropic
